apps/api/pyproject.toml
[project]
name = "raven-api"
version = "0.1.0"
description = "Raven AI Gateway - Python API"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis[hiredis]>=7.1.0",
    "cryptography>=44.0.0",
    "cuid2>=2.0.0",
    "pydantic-settings>=2.7.0",
    "litellm>=1.60.0",
    "httpx[http2]>=0.28.0",
    "fastapi-users[sqlalchemy]>=14.0.0",
    "pwdlib[argon2]>=0.2.0",
    "resend>=2.26.0",
    "aiosmtplib>=5.1.0",
    "Jinja2>=3.1.0",
    "mjml-python>=1.3.7",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = ["B008"]  # Depends() in defaults is standard FastAPI pattern

[tool.ruff.lint.isort]
known-first-party = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
